copy'n'paste mistype
git-svn-id: https://bl-toolkit.googlecode.com/svn/trunk/2@290 827d3e3c-cb35-0410-a708-6b02db8f8db3

+++ Source/Common/Convert.cs
@@ -1378,7 +1378,7 @@ namespace BLToolkit.Common
 
 			if (t == typeof(Byte?))       return (ConvertMethod)(object)(Convert<Int16?,Byte?>.ConvertMethod)      (delegate(Byte?       p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
 			if (t == typeof(UInt16?))     return (ConvertMethod)(object)(Convert<Int16?,UInt16?>.ConvertMethod)    (delegate(UInt16?     p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
-			if (t == typeof(UInt16?))     return (ConvertMethod)(object)(Convert<Int16?,UInt16?>.ConvertMethod)    (delegate(UInt16?     p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
+			if (t == typeof(UInt32?))     return (ConvertMethod)(object)(Convert<Int16?,UInt32?>.ConvertMethod)    (delegate(UInt32?     p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
 			if (t == typeof(UInt64?))     return (ConvertMethod)(object)(Convert<Int16?,UInt64?>.ConvertMethod)    (delegate(UInt64?     p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
 
 			if (t == typeof(Char?))       return (ConvertMethod)(object)(Convert<Int16?,Char?>.ConvertMethod)      (delegate(Char?       p) { return p.HasValue? (Int16?)Convert.ToInt16(p.Value) : null; });
